+++ js/widgets/card/cardrenderer.sr5.js
@@ -136,6 +136,7 @@ Object.assign(CardRenderer, {
     combatBody += this._weaponBlock(pnj, weapons, "sr5", deps);
     combatBody += this._spellsBlock(pnj, spells, "sr5");
     combatBody += this._complexFormsBlock(pnj, complexForms, "sr5");
+    combatBody += this._statusRow(pnj, "sr5", deps);
     combatBody += this._drugRow(pnj, "sr5", deps);
     combatBody += this._vehicleChipRow(pnj, deps);
     combatBody += this._spiritChipRow(pnj, deps);