+++ js/widgets/card/cardrenderer.sr6.js
@@ -136,6 +136,7 @@ Object.assign(CardRenderer, {
     combatBody += this._weaponBlock(pnj, weapons, "sr6", deps);
     combatBody += this._spellsBlock(pnj, spells, "sr6");
     combatBody += this._complexFormsBlock(pnj, complexForms, "sr6");
+    combatBody += this._statusRow(pnj, "sr6", deps);
     combatBody += this._drugRow(pnj, "sr6", deps);
     combatBody += this._vehicleChipRow(pnj, deps);
     combatBody += this._spiritChipRow(pnj, deps);